+++ include/lh/pair/fields.h
@@ -1,5 +1,5 @@
-#ifndef LH_PAIR_BOUNDS_FIELDS_H
-#define LH_PAIR_BOUNDS_FIELDS_H
+#ifndef LH_PAIR_FIELDS_H
+#define LH_PAIR_FIELDS_H
 
 /**
  * @brief A pair of typed values.
@@ -7,8 +7,8 @@
  * @param first_type Type of the first value.
  * @param second_type Type of the second value.
  */
-#define lh_pair_bounds_fields(first_type, second_type)                         \
+#define lh_pair_fields(first_type, second_type)                                \
     first_type first;                                                          \
     second_type second
 
-#endif /* LH_PAIR_BOUNDS_FIELDS_H */
+#endif /* LH_PAIR_FIELDS_H */